requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
billiard==4.2.0
cachetools==5.2.0
celery==5.4.0
certifi==2022.9.24
cffi==1.16.0
channels==3.0.4
channels-redis==4.2.0
charset-normalizer==2.1.1
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cron-descriptor==1.4.3
cryptography==42.0.8
daphne==3.0.2
Django==5.0.6
django-celery-beat==2.6.0
django-cors-headers==3.13.0
django-environ==0.9.0
django-timezone-field==6.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
google-api-core==2.10.2
google-api-python-client==2.65.0
google-auth==2.14.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.56.4
httplib2==0.21.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
kombu==5.3.7
msgpack==1.0.8
pillow==10.3.0
prompt_toolkit==3.0.47
protobuf==4.21.9
psycopg2==2.9.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.22
PyJWT==2.6.0
pyOpenSSL==24.1.0
pyparsing==3.0.9
python-crontab==3.1.0
python-dateutil==2.9.0.post0
pytz==2022.5
redis==5.0.5
requests==2.28.1
rsa==4.9
service-identity==24.1.0
six==1.16.0
sqlparse==0.4.3
stripe==9.11.0
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==1.26.12
vine==5.1.0
wcwidth==0.2.13
zope.interface==6.4.post2
